fix: card, cards 클래스 분리
